pyproject.toml
[tool.poetry]
name = "enemy"
version = "0.0.0"
description = "an adversarial discord bot"
authors = ["Kyle Moore <moore.kyle39@gmail.com>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.8.0"
"discord.py" = { version = "^1.7.3", extras = ["discord"] }

[tool.poetry.scripts]
enemy = 'enemy.client'

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}
pytest = "^7.1.0"
nox = "^2022.1.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
